nutraflex/requirements.txt
streamlit==1.29.0
google-generativeai==0.3.2
Pillow==10.0.1
streamlit-option-menu==0.3.6
python-dotenv==1.0.0